Main/django-rest/requirements.txt
django==4.2.4
djangorestframework
pyyaml
requests
django-cors-headers
